Working ajax image dashboard
The user's dashboard is initally set up in a 3 X 3 square grid. 9 images will load (including the new upload square). A button at the bottom of the page uses ajax to make a request from the server, sending it the necessary data to query the database (tracking the current index on the client side in JavaScript.) and the server returns JSON which is then looped through to create necessary HTML elements to append to the proper div. The query contains a range of 9 items each time. If there are no more items to return, the page will disable the "more images" button.

This may be done using scroll-spy in the future, to automatically load the batches of images as the user scrolls down.